Revert "fix: don't assume boundary exists during increment/decrement (#18289)"
This reverts commit b40c359d44196b7b87745259243333763ed61ec5.
