+++ packages/svelte/tests/runtime-runes/samples/async-async-disconnected-effect-root/main.svelte
@@ -1,0 +1,11 @@
+<script>
+	setTimeout(() => {
+		$effect.root(() => {
+			async function fn() {
+				const value = $derived(await 1);
+				return { get value() { return value } };
+			}
+			fn().then(r => console.log(r.value));
+		})
+	})
+</script>
