--- packages/svelte/tests/runtime-runes/samples/async-unset-context-after-restore/_config.js
@@ -1,14 +1,0 @@
-import { tick } from 'svelte';
-import { test } from '../../test';
-
-export default test({
-	async test({ assert, logs }) {
-		await new Promise((resolve) => setTimeout(resolve, 10));
-		assert.deepEqual(logs, [1, 1]);
-		const [button] = document.querySelectorAll('button');
-
-		button.click();
-		await tick();
-		assert.deepEqual(logs, [1, 1, 2]);
-	}
-});
